feat(noosphere): Introduce `noosphere` crate
The `noosphere` crate is the highest level entrypoint into Noosphere
APIs that we offer. It re-exports the functionality of most of our other
crates, and offers FFI interfaces for C and JS/WASM consumers.

In addition to the `noosphere` crate, this change introduces some major
improvements to our Github Actions workflows. Workflows now let you
build associated artifacts on demand, and are also assembled into a
larger release orchestration.