feat: Refactor storage interfaces
The `noosphere-storage` crate is cleaned up and significantly refactored
to support more ergonomic adaptation to other backend storage scenarios.

BREAKING CHANGE: The `StorageProvider` trait has been replaced by the
`Storage` trait. This new trait allows for distinct backing
implementations of `BlockStore` and `KeyValueStore`.

BREAKING CHANGE: A `SphereDb` was previously generic over a `Store`, but
is now generic over the new `Storage` trait.

BREAKING CHANGE: The modules in the `noosphere-storage` crate have been
re-arranged so that they are organized more consistently with the way we
organize modules in other crates.